refactor: WinningNumbers, BonusNumber 검증 LottoResult로 이동
